requirements.txt
fastapi==0.112.0
uvicorn==0.30.1
langchain==0.1.14
langchain-google-genai==0.0.9
google-generativeai==0.3.1
python-dotenv==1.0.1
pydantic~=2.11.6
h11~=0.16.0
pip~=25.1.1
protobuf~=4.25.8
rsa~=4.9.1
pyasn1~=0.6.1
grpcio~=1.73.0
requests~=2.32.4
PyYAML~=6.0.2
idna~=3.10
multidict~=6.4.4
propcache~=0.3.2
exceptiongroup~=1.3.0
sniffio~=1.3.1
attrs~=25.3.0
click~=8.1.8
httpcore~=1.0.9
httpx~=0.28.1
numpy~=1.26.4
setuptools~=58.0.4
charset-normalizer~=3.4.2
proto-plus~=1.26.1
aiohttp~=3.12.12
pyasn1-modules~=0.4.2
cachetools~=5.5.2
certifi~=2025.4.26
urllib3~=2.4.0
packaging~=23.2
grpcio-status~=1.62.3
tqdm~=4.67.1
orjson~=3.10.18
yarl~=1.20.1
async-timeout~=4.0.3
aiosignal~=1.3.2
frozenlist~=1.7.0
aiohappyeyeballs~=2.6.1
starlette~=0.37.2
anyio~=4.9.0
annotated-types~=0.7.0
tenacity~=8.5.0
langchain-core~=0.1.53
langchain-community~=0.0.38
langsmith~=0.1.147
mypy-extensions~=1.1.0
langchain-text-splitters~=0.0.2
sqlalchemy~=2.0.41
requests-toolbelt~=1.0.0
marshmallow~=3.26.1
jsonpointer~=3.0.0
jsonpatch~=1.33
typing-inspect~=0.9.0
dataclasses-json~=0.6.7
pytest~=8.4.0
pygments~=2.19.1